+++ index.js
@@ -1,7 +1,7 @@
 /**
  * @format
  */
-
+import 'react-native-gesture-handler'; // Importante: sempre no topo
 import {AppRegistry} from 'react-native';
 import App from './src/App';
 import {name as appName} from './app.json';
