+++ package.json
@@ -10,8 +10,16 @@
     "test": "jest"
   },
   "dependencies": {
+    "@react-navigation/bottom-tabs": "^7.0.6",
+    "@react-navigation/drawer": "^7.0.6",
+    "@react-navigation/native": "^7.0.3",
+    "@react-navigation/native-stack": "^7.1.0",
+    "@react-navigation/stack": "^7.0.5",
     "react": "18.3.1",
-    "react-native": "0.76.1"
+    "react-native": "0.76.1",
+    "react-native-gesture-handler": "^2.21.2",
+    "react-native-safe-area-context": "^4.14.0",
+    "react-native-screens": "^4.2.0"
   },
   "devDependencies": {
     "@babel/core": "^7.25.2",
