+++ src/App.tsx
@@ -1,64 +1,19 @@
-import React, { Component } from 'react';
-import {
-  SafeAreaView,
-  StyleSheet,
-  Text,
-  TextInput,
-} from 'react-native';
+import react from 'react';
+import { NavigationContainer } from '@react-navigation/native';
+import { createNativeStackNavigator } from '@react-navigation/native-stack';
 
-import { Generator } from './components/loto/Generator';
-import NumbersSelector from './modals/NumbersSelector';
-import Action from './components/button/Action';
+import Favorites from './views/Favorites';
+import Game from './views/Game';
 
-export default class App extends Component {
+const Stack = createNativeStackNavigator();
 
-  state = {
-    showModal: false,
-    amount: 6,
-  }
-
-  updateAmount = (size: string) => this.setState({ amount: +size});
-
-  render() {
+export default function App() {
     return (
-      <SafeAreaView style={style.App}>
-        <Text style={style.TextDefault}>Chose the number amount:</Text>
-        <TextInput
-          style={[style.Input, style.TextDefault]}
-          keyboardType='numeric'
-          value={`${this.state.amount}`}
-          onChangeText={this.updateAmount}
-        />
-        <NumbersSelector
-          amount={this.state.amount}
-          isVisible={this.state.showModal}
-          onCancel={() => this.setState({ showModal: false })}
-        />
-        <Action
-          label='MAKE YOURS'
-          onClick={() => this.setState({ showModal: true })}
-        />
-        <Generator amount={this.state.amount} />
-      </SafeAreaView>
+      <NavigationContainer>
+        <Stack.Navigator initialRouteName="Game">
+          <Stack.Screen name="Favorites" component={Favorites} />
+          <Stack.Screen name="Game" component={Game} />
+        </Stack.Navigator>
+      </NavigationContainer>
     );
-  }
-};
-
-const style = StyleSheet.create({
-  App: {
-    flexGrow: 1,
-    justifyContent: 'center',
-    alignItems: 'center',
-    fontSize: 12,
-    textAlign: 'justify',
-  },
-  Input: {
-    borderBottomWidth: 1,
-  },
-  TextDefault: {
-    fontSize: 16,
-    textAlign: 'center',
-    fontWeight: 'bold',
-    padding: 5,
-  },
-});
+}
