+++ src/views/Favorites.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Text, SafeAreaView } from 'react-native';
+
+const Favorites = () => {
+  return (
+    <SafeAreaView>
+      <Text>Favorites Screen</Text>
+    </SafeAreaView>
+  );
+};
+
+export default Favorites;
